refactor: move types to model folder
