+++ src/App.tsx
@@ -2,26 +2,7 @@ import React, { ChangeEvent } from 'react';
 import './App.css';
 import SearchBar from './components/SearchBar/SearchBar';
 import SearchResults from './components/SearchResults/SearchResults';
-
-interface IAppProps {}
-
-interface IAppState {
-  searchTerm: string;
-  searchResults: IGoogleBooksApiItem[];
-  error: Error | null;
-  throwError: boolean;
-}
-
-export interface IGoogleBooksApiItem {
-  id: string;
-  volumeInfo: {
-    title: string;
-    description?: string;
-    imageLinks?: {
-      thumbnail: string;
-    };
-  };
-}
+import { IAppProps, IAppState, IGoogleBooksApiItem } from './model/App';
 
 class App extends React.Component<IAppProps, IAppState> {
   public constructor(props: IAppProps) {
