+++ src/model/ErrorBoundary.ts
@@ -1,0 +1,9 @@
+import { ReactNode } from 'react';
+
+export interface IErrorBoundaryProps {
+  children: ReactNode;
+}
+
+export interface IErrorBoundaryState {
+  hasError: boolean;
+}
