+++ src/model/SearchBar.ts
@@ -1,0 +1,9 @@
+import { ChangeEvent } from 'react';
+
+interface ISearchBarProps {
+  searchTerm: string;
+  onInputChange: (event: ChangeEvent<HTMLInputElement>) => void;
+  onSearchSubmit: () => void;
+}
+
+export default ISearchBarProps;
