+++ src/model/SearchResults.ts
@@ -1,0 +1,7 @@
+import { IGoogleBooksApiItem } from './App';
+
+interface ISearchResultsProps {
+  searchResults: IGoogleBooksApiItem[];
+}
+
+export default ISearchResultsProps;
